backend/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-environ==0.11.2
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
phonenumbers==8.13.52
sqlparse==0.5.3
